feat: implement context menu in docs
